+++ ui/src/pages/chat/components/chat-tool-cards.test.ts
@@ -9,15 +9,7 @@ import {
   formatCollapsedToolSummaryText,
   resolveCollapsedToolArgumentPreview,
 } from "../../../lib/chat/tool-cards.ts";
-import {
-  allowWidgetPrompt,
-  renderToolCard,
-  renderToolPreview,
-  resolveWidgetPromptText,
-  WIDGET_PROMPT_EVENT,
-  WIDGET_PROMPT_MAX_CHARS,
-  type WidgetPromptEventDetail,
-} from "./chat-tool-cards.ts";
+import { renderToolCard, renderToolPreview } from "./chat-tool-cards.ts";
 
 function requireFirstMockArg(
   mock: ReturnType<typeof vi.fn>,
@@ -113,154 +105,6 @@ describe("tool-cards", () => {
     );
 
     expect(container.querySelector("iframe")).not.toBeNull();
-  });
-
-  it("validates widget prompt text", () => {
-    expect(resolveWidgetPromptText("  Run the tests  ")).toBe("Run the tests");
-    expect(resolveWidgetPromptText("")).toBeNull();
-    expect(resolveWidgetPromptText("   ")).toBeNull();
-    expect(resolveWidgetPromptText(42)).toBeNull();
-    // Slash commands would run UI commands such as /approve on the widget's
-    // behalf; the send path must only ever receive conversational text.
-    expect(resolveWidgetPromptText("/approve")).toBeNull();
-    expect(resolveWidgetPromptText("  /stop")).toBeNull();
-    expect(resolveWidgetPromptText("x".repeat(WIDGET_PROMPT_MAX_CHARS + 1))).toBeNull();
-    expect(resolveWidgetPromptText("x".repeat(WIDGET_PROMPT_MAX_CHARS))).not.toBeNull();
-  });
-
-  it("rate limits widget prompts per widget document", () => {
-    const src = `/__openclaw__/canvas/documents/cv_rate_${Math.random()}/index.html`;
-    const start = 1_000_000;
-    for (let index = 0; index < 10; index += 1) {
-      expect(allowWidgetPrompt(src, start + index)).toBe(true);
-    }
-    expect(allowWidgetPrompt(src, start + 10)).toBe(false);
-    // The rolling window frees budget once earlier sends age out.
-    expect(allowWidgetPrompt(src, start + 61_000)).toBe(true);
-  });
-
-  function renderWidgetPreviewFrame(url: string, allowExternalEmbedUrls = false) {
-    const container = document.createElement("div");
-    document.body.append(container);
-    render(
-      renderToolPreview(
-        {
-          kind: "canvas",
-          surface: "assistant_message",
-          render: "url",
-          viewId: "cv_prompt",
-          url,
-        },
-        "chat_message",
-        allowExternalEmbedUrls ? { allowExternalEmbedUrls } : {},
-      ),
-      container,
-    );
-    const frame = container.querySelector("iframe");
-    expect(frame).not.toBeNull();
-    expect(frame!.contentWindow).not.toBeNull();
-    return { container, frame: frame! };
-  }
-
-  function offerPromptPort(frame: HTMLIFrameElement): MessagePort {
-    const channel = new MessageChannel();
-    window.dispatchEvent(
-      new MessageEvent("message", {
-        data: { type: "openclaw:widget-prompt-offer" },
-        origin: "null",
-        source: frame.contentWindow,
-        ports: [channel.port2],
-      }),
-    );
-    return channel.port1;
-  }
-
-  const flushPorts = () => new Promise((resolve) => setTimeout(resolve, 0));
-
-  it("adopts the bridge's prompt port offer and forwards validated prompts", async () => {
-    const { container, frame } = renderWidgetPreviewFrame(
-      "/__openclaw__/canvas/documents/cv_prompt/index.html",
-    );
-    // The bridge posts its offer at parse time, before the frame's load event.
-    const port = offerPromptPort(frame);
-    frame.dispatchEvent(new Event("load"));
-
-    // jsdom has no layout and cannot focus iframes; emulate a visible frame
-    // the user clicked into, which is what the host checks require.
-    (frame as HTMLIFrameElement & { checkVisibility: () => boolean }).checkVisibility = () => true;
-    Object.defineProperty(document, "activeElement", { get: () => frame, configurable: true });
-
-    const received: string[] = [];
-    container.addEventListener(WIDGET_PROMPT_EVENT, (event) => {
-      received.push((event as CustomEvent<WidgetPromptEventDetail>).detail.text);
-    });
-    try {
-      port.postMessage({ type: "openclaw:widget-prompt", prompt: "  Show details  " });
-      await flushPorts();
-      expect(received).toEqual(["Show details"]);
-      // Slash commands never reach the send path.
-      port.postMessage({ type: "openclaw:widget-prompt", prompt: "/approve" });
-      await flushPorts();
-      expect(received).toEqual(["Show details"]);
-      // A replacement document's later offer must not displace or re-arm the
-      // adopted grant, even across another load event.
-      const lateOfferPort = offerPromptPort(frame);
-      frame.dispatchEvent(new Event("load"));
-      lateOfferPort.postMessage({ type: "openclaw:widget-prompt", prompt: "From takeover" });
-      await flushPorts();
-      expect(received).toEqual(["Show details"]);
-      // Without focus on the frame there is no user-activation signal; drop.
-      delete (document as unknown as Record<string, unknown>).activeElement;
-      port.postMessage({ type: "openclaw:widget-prompt", prompt: "Auto send" });
-      await flushPorts();
-      expect(received).toEqual(["Show details"]);
-    } finally {
-      delete (document as unknown as Record<string, unknown>).activeElement;
-      container.remove();
-    }
-  });
-
-  it("adopts a prompt offer that arrives after the frame's load event", async () => {
-    const { container, frame } = renderWidgetPreviewFrame(
-      "/__openclaw__/canvas/documents/cv_late_offer/index.html",
-    );
-    // Posted-message and load tasks have no guaranteed ordering; here load wins.
-    frame.dispatchEvent(new Event("load"));
-    const port = offerPromptPort(frame);
-    (frame as HTMLIFrameElement & { checkVisibility: () => boolean }).checkVisibility = () => true;
-    Object.defineProperty(document, "activeElement", { get: () => frame, configurable: true });
-    const received: string[] = [];
-    container.addEventListener(WIDGET_PROMPT_EVENT, (event) => {
-      received.push((event as CustomEvent<WidgetPromptEventDetail>).detail.text);
-    });
-    try {
-      port.postMessage({ type: "openclaw:widget-prompt", prompt: "Late but valid" });
-      await flushPorts();
-      expect(received).toEqual(["Late but valid"]);
-    } finally {
-      delete (document as unknown as Record<string, unknown>).activeElement;
-      container.remove();
-    }
-  });
-
-  it("never adopts prompt offers from externally allowed embed URLs", async () => {
-    const { container, frame } = renderWidgetPreviewFrame("https://canvas.example/widget", true);
-    const port = offerPromptPort(frame);
-    frame.dispatchEvent(new Event("load"));
-    (frame as HTMLIFrameElement & { checkVisibility: () => boolean }).checkVisibility = () => true;
-    Object.defineProperty(document, "activeElement", { get: () => frame, configurable: true });
-    const received: string[] = [];
-    container.addEventListener(WIDGET_PROMPT_EVENT, (event) => {
-      received.push((event as CustomEvent<WidgetPromptEventDetail>).detail.text);
-    });
-    try {
-      port.postMessage({ type: "openclaw:widget-prompt", prompt: "External send" });
-      await flushPorts();
-      expect(received).toEqual([]);
-    } finally {
-      delete (document as unknown as Record<string, unknown>).activeElement;
-      container.remove();
-    }
   });
 
   it("keeps selected summary text from toggling the disclosure", () => {
